# backend/docs/swagger.yaml
basePath: /api
definitions:
  auth.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        description: opcional — usuário pode cadastrar sem telefone
        type: string
    required:
    - email
    - password
    type: object
  books.criarLivroRequest:
    properties:
      author:
        type: string
      colorIdx:
        type: integer
      genres:
        items:
          type: string
        type: array
      rating:
        type: integer
      review:
        type: string
      title:
        type: string
    type: object
  courses.cursoRequest:
    properties:
      certificate:
        type: boolean
      certificate_expiry:
        type: string
      credential:
        type: string
      duration:
        type: string
      end_date:
        type: string
      platform:
        type: string
      progress:
        type: integer
      start_date:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
host: localhost:3001
info:
  contact: {}
  description: 'API do BeePlanner — autenticação via JWT (Bearer). Todas as rotas
    /api/* (exceto /auth) exigem o header Authorization: Bearer <token>.'
  title: BeePlanner API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciais
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - Auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados de cadastro
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cadastrar usuário
      tags:
      - Auth
  /books:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar livros
      tags:
      - Books
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do livro
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/books.criarLivroRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar livro
      tags:
      - Books
  /books/{id}:
    delete:
      parameters:
      - description: ID do livro
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deletar livro
      tags:
      - Books
  /courses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar cursos
      tags:
      - Courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do curso
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/courses.cursoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar curso
      tags:
      - Courses
  /courses/{id}:
    delete:
      parameters:
      - description: ID do curso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Deletar curso
      tags:
      - Courses
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do curso
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/courses.cursoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Editar curso
      tags:
      - Courses
...
securityDefinitions:
  BearerAuth:
    description: 'Formato: Bearer <seu_jwt>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"